Поменял структуру файлов, добавил докерфайлы для streamlit приложения и сервера, написал docker-compose
