+++ docker-compose.yml
@@ -1,0 +1,29 @@
+version: '3'
+
+services:
+  backend:
+    container_name: backend
+    build: 
+      context: .
+      dockerfile: backend/Dockerfile
+    environment:
+      BACKEND_PORT: "8080"
+    restart: unless-stopped
+    healthcheck:
+      test: ["CMD", "curl", "-f", "http://backend:8080/api/v1/healthcheck"]
+      interval: 20s
+      timeout: 20s
+      retries: 5
+    ports:
+      - 8080:8080
+
+  frontend:
+    container_name: frontend
+    environment:
+      BACKEND_URI: backend
+    build: 
+      context: .
+      dockerfile: streamlit/Dockerfile
+    restart: unless-stopped
+    ports:
+      - 8501:8501
