+++ streamlit/resources/EDA.md
@@ -1,0 +1,1 @@
+Проведенный разведочный анализ адресов банкоматов показал, что имеющиеся адреса представлены в муниципальном формате. Данный формат дает государству лучше понять, каким органам управления подчиняется адрес и бизнесу легче понимать какому отделению налоговой он должен сдавать отчетность. В муниципальном делении указывают структуры местного самоуправления, например, регион, районы, сельские и городские поселения, муниципальные округа. Данный факт усложняет формат записи адреса для парсинга, поскольку адреса могут иметь самые различные составляющие. Также в данных имеются записи о банкоматах в нашем датасете, адресов на русском с географическими координатами которых нет вовсе.
